requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pytest
pytest-cov